+++ CHANGELOG.md
@@ -2,9 +2,13 @@
 
 ## Release 1.4.2
 
+- [Fix] Fixed pagination issue
 - [Feat] Added new query methods
 - [Refactor] Refactored Query builders
+- [Refactor] Refactored posts Collections to allow any [Collection](https://laravel.com/docs/8.x/collections) method
+- [Refactor] `Timber::get_posts()` refactored to `new \Timber\PostQuery()` as it was needed for pagination
 - [Refactor] Refactored `query.php` configuration file - added `limit` option which set `posts_per_page` ONLY within all method
+- [Docs] Changed return type for `Route` facade to help IDE
 
 Release Date: Jan 20th, 2022